LayoutTests/streams/readable-stream-getReader-expected.txt


PASS getReader() should return a ReadableStreamDefaultReader
PASS getReader({mode: 'byob'}) should return a TypeError as BYOB reader is not yet implemented
PASS getReader({mode: 'aMode'}) should return a RangeError as 'aMode' is different from undefined and 'byob'
